DLUHC-204 move PDP api calls into sub folder
